Django3 compatibility
I noticed that Django3 has an immutable QueryDict,
so if we try to update it, it will fail. Making a
copy and forcing it to be a dict will solved these
problems (tested with djangosaml2idp as IDP and
another djangosaml2idp as SP, both SP-initiated
as well as IDP-initiated logins).