Improved PersistentId support to not depend on hash of email
A user email, service provider url, or identity provider url might change over time. In the previous system, which relies on all these things, this would change the PersistendId, leading to possible account loss on the client side. By using a random uuid stored in the database, this risk is alleviated.